Extract components for ‘Stadia’ section
